backend/requirements.txt

fastapi
uvicorn
langchain
langchain-openai
langgraph
openai
pinecone-client
boto3
pymupdf
python-dotenv
